jupyter-kill-kernel (jupyter-server-kernel): Raise an error
The purpose of this method was mainly to kill the kernel process if a clean
shutdown is not possible, but this is already done on the sever side. Also, the
previous behavior of making the ID slot of a kernel unbound doesn't work well
since the `jupyter-event-handler` of a `jupyter-server` relies on the ID slot.
Messages could still be coming in when `jupyter-kill-kernel` is called.